// homework_07/src/api/JsonPlaceholderAPI/types.ts
export interface FetchArgs<Updates = Record<string, unknown>> {
  signal: AbortSignal
  userId?: number
  // updates?:  Record<string, unknown>
  updates?: Updates
  pageNumber?: number
  postId?: number
}

export interface Photo {
  albumId: number
  id: number
  title: string
  url: string
  thumbnailUrl: string
}

export interface User {
  id: number
  name: string
  username: string
  email: string
  address: Address
  phone: string
  website: string
  company: Company
}

export interface Address {
  street: string
  suite: string
  city: string
  zipcode: string
  geo: Geo
}

export interface Geo {
  lat: string
  lng: string
}

export interface Company {
  name: string
  catchPhrase: string
  bs: string
}

export interface Post {
  userId: number
  id: number
  title: string
  body: string
  userName?: string
}

export interface Comment {
  postId: number
  id: number
  name: string
  email: string
  body: string
}